fix handling of containers in resolve_registry_includes (fixes #12)
The containers' source_file_line points to the template's
definition, which is then resolved to whichever header
included it first. This commit adds a special case to ensure
that the metadata is properly set

Note that it does not remove the corresponding hardcoded include values
in include_for_type as it opens a can of worms during opaque generation.
In this phase, we create our own types (in particular vectors-of-opaques)
for which the normal load mechanisms do not apply - and therefore for
which the orogen_include metadata is set *after* the current generation
pass needed it:

  - create m-types < this requires the orogen_include info
  - load m-types < this creates the orogen_include info

It can be fixed, but requires quite a bit of changes, so delay that
until we manage to make the Typekit class something looking less horrid